codebuddy/rag_ingest/pyproject.toml
[project]
name = "rag_ingest"
version = "0.1.0"
description = "rag service"
readme = "README.md"
authors = [
    { name = "Xudong Zhang", email = "zhangxudong@lixiang.com" }
]
requires-python = ">=3.10"
dependencies = [
    "llama-index-core>=0.11.18",
    "tree-sitter<0.22",
    "tree-sitter-go>=0.23.3",
    "tree-sitter-languages>=1.10.2",
    "openai>=1.54.4",
    "psycopg2>=2.9.10",
    "fire>=0.7.0",
    "jinja2>=3.1.4",
    "sysutils",
    "repoutils",
    "aiohttp>=3.11.11",
]

[[tool.uv.index]]
# url = "https://mirrors.aliyun.com/pypi/simple"
url = "https://artifactory.ep.chehejia.com/artifactory/api/pypi/pypi-remote/simple"
default = true

[tool.uv.sources]
sysutils = { path = "../../packages/sysutils", editable = true }
repoutils = { path = "../../packages/repoutils", editable = true }


[tool.pytest.ini_options]
log_cli = "true"
log_cli_level = "DEBUG"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ['rag_ingest']
